Fixing the cfg-gating for futures
