+++ src/lib.rs
@@ -94,7 +94,7 @@ extern crate stdweb_internal_macros;
 #[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
 pub use stdweb_internal_macros::js_export;
 
-#[cfg(any(test, feature = "futures"))]
+#[cfg(feature = "futures")]
 extern crate futures;
 
 #[macro_use]
@@ -131,7 +131,10 @@ pub use webcore::instance_of::InstanceOf;
 pub use webcore::reference_type::ReferenceType;
 pub use webcore::serialization::JsSerialize;
 
-pub use webcore::promise::{Promise, PromiseFuture};
+pub use webcore::promise::Promise;
+
+#[cfg(feature = "futures")]
+pub use webcore::promise_future::PromiseFuture;
 
 #[cfg(feature = "serde")]
 /// A module with serde-related APIs.
