+++ src/webcore/mod.rs
@@ -16,6 +16,11 @@ pub mod once;
 pub mod instance_of;
 pub mod reference_type;
 pub mod promise;
+
+#[cfg(feature = "futures")]
+pub mod promise_future;
+
+#[cfg(feature = "futures")]
 pub mod promise_executor;
 
 #[cfg(feature = "nightly")]
